Auto Enroll

+++ src/Tools/_framework/ToolPanels/SignInRedirector.jsx
@@ -7,11 +7,15 @@ import { pageToolViewAtom, searchParamAtomFamily } from '../NewToolRoot';
 export default function SignInRedirector() {
   const doenetId = useRecoilValue(searchParamAtomFamily('doenetId'));
   const setPageToolView = useSetRecoilState(pageToolViewAtom);
-  
-      axios.get(`/api/umn/shibToJWT.php`)
+
+      axios.get(`/api/umn/shibToJWT.php`,{params:{doenetId}})
       .then(({data})=>{
         console.log("data",data);
         if (data.success){
+          if (!data.isEnrolled){
+            //SEND ERROR!!!
+            return null;
+          }
           if (data.needToClearOutPreviousUser){
             localStorage.clear(); //Clear out the profile
             indexedDB.deleteDatabase('keyval-store'); //Clear out the rest of the profile
@@ -19,7 +23,7 @@ export default function SignInRedirector() {
           //Redirect to exam
           setPageToolView(
              {
-              page:"placementExam",
+              page:"placementexam",
               tool: 'exam',
               view: '',
               params: {
